+++ ts/src/render-registrations.ts
@@ -5,3 +5,4 @@
 import "./markup/render";
 import "./layout/render";
 import "./rich-text/render";
+import "./download-card/render";